pyproject.toml
[project]
name = "voxgate"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "httpx>=0.25",
    "anthropic>=0.40",
    "google-auth>=2.30",
    "itsdangerous>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "httpx>=0.25",
    "ruff>=0.4",
]

[tool.setuptools]
py-modules = ["server"]
packages = ["auth"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
